requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
chardet==4.0.0
disnake==2.1.2
idna==3.2
multidict==5.2.0
python-dotenv==0.19.0
typing-extensions==3.10.0.2
yarl==1.7.0
